requirements.txt
# GitHub integration
PyGithub==1.59.0

# Date handling
python-dateutil==2.8.2
tkcalendar>=1.6.1

# GUI
python-tk; platform_system=="Linux"